# docs/swagger.yaml
basePath: /
definitions:
  github_com_noireveil_ecoserve-backend_internal_domain.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      otpcode:
        type: string
      otpexpiresAt:
        type: string
      profilePictureURL:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_noireveil_ecoserve-backend_internal_usecase.CompleteOrderRequest:
    properties:
      category:
        type: string
      device_weight:
        type: number
      distance_km:
        type: number
      latitude:
        type: number
      longitude:
        type: number
      photo_url:
        type: string
      service_fee:
        type: number
    type: object
  github_com_noireveil_ecoserve-backend_internal_usecase.UpdateProfileRequest:
    properties:
      full_name:
        example: EcoServe Tester
        type: string
      profile_picture_url:
        example: https://storage.com/photo.jpg
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_delivery_http_handlers.CreateDevicePayload:
    properties:
      brand_name:
        example: Apple iPhone 14 Pro
        type: string
      category:
        example: Smartphone
        type: string
      weight_in_kg:
        example: 0.21
        type: number
    type: object
  internal_delivery_http_handlers.CreateOrderPayload:
    properties:
      customer_latitude:
        type: number
      customer_longitude:
        type: number
      device_category:
        example: Pendingin & Komersial
        type: string
      device_id:
        example: (Opsional) uuid-device-dpp
        type: string
      problem_description:
        example: Kompresor mati dan berasap
        type: string
      technician_id:
        example: (Opsional) uuid-teknisi
        type: string
    type: object
  internal_delivery_http_handlers.RegisterTechnicianRequest:
    properties:
      createdAt:
        type: string
      experienceYears:
        type: integer
      id:
        type: string
      latitude:
        example: -6.1944
        type: number
      longitude:
        example: 106.8229
        type: number
      rating:
        type: number
      specialization:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_noireveil_ecoserve-backend_internal_domain.User'
      userID:
        type: string
    type: object
  internal_delivery_http_handlers.RequestOTPPayload:
    properties:
      email:
        example: eco.security@ecoserve.com
        type: string
      full_name:
        example: EcoServe Tester
        type: string
    type: object
  internal_delivery_http_handlers.ReviewRequestPayload:
    properties:
      comment:
        example: Teknisi sangat profesional dan jujur!
        type: string
      rating:
        example: 5
        type: integer
    type: object
  internal_delivery_http_handlers.TriageRequest:
    properties:
      latitude:
        example: -6.1944
        type: number
      longitude:
        example: 106.8229
        type: number
      message:
        example: Kompresor kulkas saya mati dan bau hangus.
        type: string
    type: object
  internal_delivery_http_handlers.VerifyOTPPayload:
    properties:
      code:
        example: "123456"
        type: string
      email:
        example: ecosecurity@ecoserve.com
        type: string
    type: object
info:
  contact: {}
  description: REST API untuk Platform Manajemen Siklus Hidup Elektronik & Servis
    Sirkular.
  title: EcoServe API
  version: "1.0"
paths:
  /api/chatbot/triage:
    post:
      consumes:
      - application/json
      description: Menganalisis kerusakan dan memberikan panduan atau merujuk ke teknisi
        terdekat berdasarkan Confidence Gate.
      parameters:
      - description: Data Keluhan dan Geospasial
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handlers.TriageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Triase Kerusakan via AI
      tags:
      - Chatbot
  /api/devices/:
    get:
      description: Mengambil daftar semua perangkat (DPP) yang dimiliki oleh pengguna.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mengambil Garasi Perangkat
      tags:
      - Devices
    post:
      consumes:
      - application/json
      description: Menambahkan perangkat elektronik baru ke garasi pengguna untuk
        pembuatan Digital Product Passport.
      parameters:
      - description: Data Perangkat
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handlers.CreateDevicePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendaftarkan Perangkat (DPP)
      tags:
      - Devices
  /api/devices/{id}:
    delete:
      description: Menghapus perangkat elektronik dari garasi pengguna.
      parameters:
      - description: ID Perangkat (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Menghapus Perangkat (DPP)
      tags:
      - Devices
    get:
      description: Mengambil informasi lengkap satu perangkat berdasarkan UUID.
      parameters:
      - description: ID Perangkat (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Detail Perangkat (DPP)
      tags:
      - Devices
  /api/orders/:
    get:
      description: Mengambil daftar pesanan yang terkait dengan pengguna (sebagai
        Konsumen atau Teknisi).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Riwayat Pesanan
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Menginisiasi pesanan servis elektronik baru oleh Konsumen. Bisa
        diikat langsung ke teknisi atau dibiarkan kosong untuk pesanan publik.
      parameters:
      - description: Data Kerusakan dan Booking Opsional
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handlers.CreateOrderPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Membuat Pesanan Perbaikan
      tags:
      - Orders
  /api/orders/{id}:
    get:
      description: Mengambil informasi lengkap satu pesanan berdasarkan UUID.
      parameters:
      - description: ID Pesanan (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Detail Pesanan
      tags:
      - Orders
  /api/orders/{id}/accept:
    put:
      description: Teknisi mengambil pesanan yang berstatus PENDING (Merubah status
        menjadi ACCEPTED).
      parameters:
      - description: ID Pesanan (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Menerima Pesanan Masuk
      tags:
      - Orders
  /api/orders/{id}/complete:
    put:
      consumes:
      - application/json
      description: Menyelesaikan pesanan dengan memvalidasi foto bukti dan koordinat
        GPS (Metrik EPA WARM).
      parameters:
      - description: ID Pesanan (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Penyelesaian
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_noireveil_ecoserve-backend_internal_usecase.CompleteOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Menyelesaikan Transaksi & Kalkulasi Emisi (Anti-Fraud)
      tags:
      - Orders
  /api/orders/incoming:
    get:
      description: Mengambil daftar pesanan yang berstatus PENDING dan belum memiliki
        teknisi.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Pesanan Masuk
      tags:
      - Orders
  /api/reviews/order/{order_id}:
    post:
      consumes:
      - application/json
      description: Konsumen memberikan rating (1-5) dan komentar untuk teknisi setelah
        pesanan selesai. Sistem akan otomatis menghitung ulang rata-rata rating teknisi
        dengan fitur keamanan Anti-Spam.
      parameters:
      - description: ID Pesanan (UUID)
        in: path
        name: order_id
        required: true
        type: string
      - description: Data Ulasan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handlers.ReviewRequestPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Berhasil memberikan ulasan
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error validasi (rating tidak valid, pesanan belum selesai,
            dll)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Akses ditolak (belum login atau token tidak valid)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Memberikan Ulasan Teknisi
      tags:
      - Reviews
  /api/technicians/:
    post:
      consumes:
      - application/json
      description: Menambahkan data teknisi baru beserta titik koordinat operasinya.
      parameters:
      - description: Data Teknisi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handlers.RegisterTechnicianRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Mendaftarkan Teknisi Baru
      tags:
      - Technicians
  /api/technicians/earnings:
    get:
      description: Mengambil total pendapatan dan pendapatan bulan ini berdasarkan
        akumulasi tarif jasa pesanan yang telah selesai.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Pendapatan Teknisi
      tags:
      - Technicians
  /api/technicians/nearby:
    get:
      description: Melakukan kueri spasial (PostGIS) untuk mencari teknisi dalam radius
        tertentu.
      parameters:
      - description: Garis Bujur (Longitude)
        in: query
        name: lon
        required: true
        type: number
      - description: Garis Lintang (Latitude)
        in: query
        name: lat
        required: true
        type: number
      - description: 'Radius pencarian dalam KM (Default: 10)'
        in: query
        name: radius_km
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Cari Teknisi Terdekat
      tags:
      - Technicians
  /api/technicians/performance:
    get:
      description: Mengambil agregasi data performa (Rating, Total Perbaikan, dan
        Total CO2 yang diselamatkan) dari teknisi yang sedang login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Metrik Performa Teknisi
      tags:
      - Technicians
  /api/users/auth/logout:
    post:
      description: Mencabut sesi pengguna dengan menghapus cookie JWT dari browser.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Logout Pengguna
      tags:
      - Auth
  /api/users/auth/request:
    post:
      consumes:
      - application/json
      description: Mengirimkan kode OTP ke email pengguna untuk proses otentikasi.
      parameters:
      - description: Data Pengguna
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handlers.RequestOTPPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Meminta Kode OTP
      tags:
      - Auth
  /api/users/auth/verify:
    post:
      consumes:
      - application/json
      description: Memverifikasi OTP dan mengeluarkan token JWT di dalam Cookie (Sesi
        Aktif).
      parameters:
      - description: Data OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handlers.VerifyOTPPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Verifikasi Kode OTP
      tags:
      - Auth
  /api/users/me:
    delete:
      description: Menghapus akun dan mencabut sesi (logout). Data tidak dihapus permanen
        untuk menjaga integritas relasi EPA WARM.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Menghapus Akun Pengguna (Soft Delete)
      tags:
      - Users
    get:
      description: Mengambil data profil dari pengguna yang sedang login berdasarkan
        JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Profil Pengguna
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Mengubah nama lengkap dan foto profil pengguna. Email tidak dapat
        diubah untuk menjaga integritas OTP.
      parameters:
      - description: Data Profil Baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_noireveil_ecoserve-backend_internal_usecase.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Memperbarui Profil Pengguna
      tags:
      - Users
  /api/users/me/impact:
    get:
      description: Mengambil agregasi data dampak lingkungan (Total Perbaikan dan
        Total CO2 yang Dihindari) untuk konsumen yang sedang login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mendapatkan Dampak Lingkungan Pengguna
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: jwt
    type: apiKey
swagger: "2.0"