# internal/delivery/http/handlers/order_handler.go
package handlers

import (
	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
	"github.com/noireveil/ecoserve-backend/internal/delivery/http/middleware"
	"github.com/noireveil/ecoserve-backend/internal/domain"
	"github.com/noireveil/ecoserve-backend/internal/usecase"
)

type OrderHandler struct {
	orderUsecase usecase.OrderUsecase
}

type CreateOrderPayload struct {
	TechnicianID       string  `json:"technician_id" example:"(Opsional) uuid-teknisi"`
	DeviceID           string  `json:"device_id" example:"(Opsional) uuid-device-dpp"`
	DeviceCategory     string  `json:"device_category" example:"Pendingin & Komersial"`
	ProblemDescription string  `json:"problem_description" example:"Kompresor mati dan berasap"`
	CustomerLatitude   float64 `json:"customer_latitude"`
	CustomerLongitude  float64 `json:"customer_longitude"`
}

func NewOrderHandler(app *fiber.App, usecase usecase.OrderUsecase) {
	handler := &OrderHandler{orderUsecase: usecase}

	api := app.Group("/api/orders")
	api.Post("/", middleware.Protected(), handler.Create)
	api.Get("/", middleware.Protected(), handler.GetMyOrders)
	api.Get("/incoming", middleware.Protected(), handler.GetIncomingOrders)
	api.Get("/:id", middleware.Protected(), handler.GetDetail)
	api.Put("/:id/accept", middleware.Protected(), handler.Accept)
	api.Put("/:id/complete", middleware.Protected(), handler.Complete)
}

// @Summary Mendapatkan Detail Pesanan
// @Description Mengambil informasi lengkap satu pesanan berdasarkan UUID.
// @Tags Orders
// @Produce json
// @Security ApiKeyAuth
// @Param id path string true "ID Pesanan (UUID)"
// @Success 200 {object} map[string]interface{}
// @Router /api/orders/{id} [get]
func (h *OrderHandler) GetDetail(c *fiber.Ctx) error {
	orderID := c.Params("id")
	order, err := h.orderUsecase.GetOrderByID(orderID)
	if err != nil {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{"error": "Pesanan tidak ditemukan"})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"message": "Berhasil mengambil detail pesanan",
		"data":    order,
	})
}

// @Summary Mendapatkan Pesanan Masuk
// @Description Mengambil daftar pesanan yang berstatus PENDING dan belum memiliki teknisi.
// @Tags Orders
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} map[string]interface{}
// @Router /api/orders/incoming [get]
func (h *OrderHandler) GetIncomingOrders(c *fiber.Ctx) error {
	orders, err := h.orderUsecase.GetIncomingOrders()
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Gagal mengambil data pesanan masuk"})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"message": "Berhasil mengambil daftar pesanan masuk",
		"data":    orders,
	})
}

// @Summary Mendapatkan Riwayat Pesanan
// @Description Mengambil daftar pesanan yang terkait dengan pengguna (sebagai Konsumen atau Teknisi).
// @Tags Orders
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} map[string]interface{}
// @Router /api/orders/ [get]
func (h *OrderHandler) GetMyOrders(c *fiber.Ctx) error {
	userIDStr, ok := c.Locals("user_id").(string)
	if !ok {
		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "User ID tidak ditemukan pada token"})
	}

	orders, err := h.orderUsecase.GetUserOrders(userIDStr)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Gagal mengambil data pesanan"})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"message": "Berhasil mengambil riwayat pesanan",
		"data":    orders,
	})
}

// @Summary Membuat Pesanan Perbaikan
// @Description Menginisiasi pesanan servis elektronik baru oleh Konsumen. Bisa diikat langsung ke teknisi atau dibiarkan kosong untuk pesanan publik.
// @Tags Orders
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param request body CreateOrderPayload true "Data Kerusakan dan Booking Opsional"
// @Success 201 {object} map[string]interface{}
// @Router /api/orders/ [post]
func (h *OrderHandler) Create(c *fiber.Ctx) error {
	var req CreateOrderPayload

	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Format JSON tidak valid"})
	}

	userIDStr, ok := c.Locals("user_id").(string)
	if !ok {
		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "User ID tidak ditemukan pada token"})
	}

	customerID, err := uuid.Parse(userIDStr)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Format User ID tidak valid"})
	}

	var techIDPtr *uuid.UUID
	if req.TechnicianID != "" {
		tID, errParse := uuid.Parse(req.TechnicianID)
		if errParse == nil {
			techIDPtr = &tID
		}
	}

	var devIDPtr *uuid.UUID
	if req.DeviceID != "" {
		dID, errParse := uuid.Parse(req.DeviceID)
		if errParse == nil {
			devIDPtr = &dID
		}
	}

	order := domain.Order{
		CustomerID:         customerID,
		TechnicianID:       techIDPtr,
		DeviceID:           devIDPtr,
		DeviceCategory:     req.DeviceCategory,
		ProblemDescription: req.ProblemDescription,
		CustomerLatitude:   req.CustomerLatitude,
		CustomerLongitude:  req.CustomerLongitude,
	}

	if err := h.orderUsecase.CreateOrder(&order); err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": err.Error()})
	}

	return c.Status(fiber.StatusCreated).JSON(fiber.Map{
		"message": "Pesanan layanan berhasil dibuat",
		"data":    order,
	})
}

// @Summary Menerima Pesanan Masuk
// @Description Teknisi mengambil pesanan yang berstatus PENDING (Merubah status menjadi ACCEPTED).
// @Tags Orders
// @Produce json
// @Security ApiKeyAuth
// @Param id path string true "ID Pesanan (UUID)"
// @Success 200 {object} map[string]interface{}
// @Router /api/orders/{id}/accept [put]
func (h *OrderHandler) Accept(c *fiber.Ctx) error {
	orderID := c.Params("id")

	userIDStr, ok := c.Locals("user_id").(string)
	if !ok {
		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "User ID tidak valid"})
	}

	role, _ := c.Locals("role").(string)
	if role != "technician" {
		return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "Hanya teknisi yang berhak mengambil pesanan"})
	}

	if err := h.orderUsecase.AcceptOrder(orderID, userIDStr); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": err.Error()})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"message": "Pesanan berhasil diambil, silakan menuju ke lokasi konsumen.",
	})
}

// @Summary Menyelesaikan Transaksi & Kalkulasi Emisi (Anti-Fraud)
// @Description Menyelesaikan pesanan dengan memvalidasi foto bukti dan koordinat GPS (Metrik EPA WARM).
// @Tags Orders
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param id path string true "ID Pesanan (UUID)"
// @Param request body usecase.CompleteOrderRequest true "Data Penyelesaian"
// @Success 200 {object} map[string]interface{}
// @Router /api/orders/{id}/complete [put]
func (h *OrderHandler) Complete(c *fiber.Ctx) error {
	orderID := c.Params("id")

	var req usecase.CompleteOrderRequest

	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Format JSON tidak valid"})
	}

	if err := h.orderUsecase.CompleteOrder(orderID, req); err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": err.Error()})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"message": "Transaksi berhasil diselesaikan. Protokol Anti-Fraud tersimpan dan Kalkulasi EPA WARM telah diinjeksikan.",
	})
}